Add Loader spinner styles and Students loader action for index
