+++ src/constants/api.js
@@ -1,3 +1,4 @@
 export const LOGIN_URL = '/api/v1/auth/sign_in';
 export const LOGOUT_URL = 'api/v1/auth/sign_out';
 export const PROFILE_URL = '/api/v1/users/profile';
+export const STUDENT_URL = '/api/v1/students'
