+++ src/index.jsx
@@ -4,25 +4,27 @@ import { createRoot } from "react-dom/client";
 import { BrowserRouter, Navigate, Outlet, Route, Routes } from "react-router-dom";
 import { RecoilRoot } from "recoil";
 import LoaderSpinner from "./jsx/components/LoaderSpinner/LoaderSpinner";
-import Error404 from './jsx/pages/Error404';
-import LoginPage from './jsx/pages/Login';
-import DashboardLayout from "./jsx/layouts/DashboardLayout";
 import ThemeContextProvider from "./context/ThemeContext";
-import Students from "./jsx/components/Student/Students";
-import StudentDetails from "./jsx/components/Student/StudentDetails";
-import AddNewStudent from "./jsx/components/Student/AddNewStudent";
-import Teachers from "./jsx/components/Teacher/Teachers";
-import TeacherDetail from "./jsx/components/Teacher/TeachersDetail";
-import AddNewTeacher from "./jsx/components/Teacher/AddNewTeacher";
-import AppProfile from "./jsx/components/AppsMenu/AppProfile/AppProfile";
-import EditProfile from "./jsx/components/AppsMenu/AppProfile/EditProfile";
-import Finance from "./jsx/components/Dashboard/Finance";
-import Home from "./jsx/components/Dashboard/Home";
+
+const Error404 = React.lazy(() => import('./jsx/pages/Error404'));
+const LoginPage = React.lazy(() => import('./jsx/pages/Login'));
+const DashboardLayout = React.lazy(() => import("./jsx/layouts/DashboardLayout"));
+const Students = React.lazy(() => import("./jsx/components/Student/Students"));
+const StudentDetails = React.lazy(() => import("./jsx/components/Student/StudentDetails"));
+const AddNewStudent = React.lazy(() => import("./jsx/components/Student/AddNewStudent"));
+const Teachers = React.lazy(() => import("./jsx/components/Teacher/Teachers"));
+const TeacherDetail = React.lazy(() => import("./jsx/components/Teacher/TeachersDetail"));
+const AddNewTeacher = React.lazy(() => import("./jsx/components/Teacher/AddNewTeacher"));
+const AppProfile = React.lazy(() => import("./jsx/components/AppsMenu/AppProfile/AppProfile"));
+const EditProfile = React.lazy(() => import("./jsx/components/AppsMenu/AppProfile/EditProfile"));
+const Finance = React.lazy(() => import("./jsx/components/Dashboard/Finance"));
+const Home = React.lazy(() => import("./jsx/components/Dashboard/Home"));
 
 
 /// Style
 import './other/swiper/css/swiper-bundle.min.css';
 import "./other/bootstrap-select/dist/css/bootstrap-select.min.css";
+import 'spinkit/spinkit.min.css'
 import "./css/style.css";
 
 const root = createRoot(document.getElementById('root'));
