+++ src/jsx/components/LoaderSpinner/LoaderSpinner.jsx
@@ -1,9 +1,15 @@
 const LoaderSpinner = () => (
   <div id="preloader">
-    <div className="sk-three-bounce">
-      <div className="sk-child sk-bounce1"></div>
-      <div className="sk-child sk-bounce2"></div>
-      <div className="sk-child sk-bounce3"></div>
+    <div class="sk-grid sk-primary">
+      <div class="sk-grid-cube bg-primary"></div>
+      <div class="sk-grid-cube bg-primary"></div>
+      <div class="sk-grid-cube bg-primary"></div>
+      <div class="sk-grid-cube bg-primary"></div>
+      <div class="sk-grid-cube bg-primary"></div>
+      <div class="sk-grid-cube bg-primary"></div>
+      <div class="sk-grid-cube bg-primary"></div>
+      <div class="sk-grid-cube bg-primary"></div>
+      <div class="sk-grid-cube bg-primary"></div>
     </div>
   </div>
 )
