+++ src/store/atoms/studentsAtom.js
@@ -1,0 +1,8 @@
+import { atom } from "recoil";
+
+const studentsAtom = atom({
+  key: 'students',
+  default:{ students: [], pagination: {}}
+})
+
+export default studentsAtom;
